Moving to ds18b20 as tempsensor since sensor sensor is not good
