+++ platformio.ini
@@ -42,7 +42,6 @@ lib_deps_gravmon =
 	; Using local copy of these libraries
 	; https://github.com/mp-se/i2cdevlib.git#<document> 		
 	; https://github.com/mp-se/Arduino-Temperature-Control-Library
-	https://github.com/pstolarz/OneWireNg#0.13.3
 	https://github.com/mp-se/tinyexpr#v1.0.0
 lib_deps_pressmon =
 	https://github.com/wollewald/ADS1115_WE#fe08755
@@ -51,6 +50,7 @@ lib_deps =
     https://github.com/256dpi/arduino-mqtt#v2.5.2
 	https://github.com/mp-se/espframework#1.1.3
 	https://github.com/mathieucarbou/ESPAsyncWebServer#v3.6.0
+	https://github.com/pstolarz/OneWireNg#0.13.3
 lib_deps8266 = 
 	https://github.com/esphome/ESPAsyncTCP#v2.0.0
 lib_deps32 = 
@@ -346,7 +346,6 @@ lib_deps =
 	${common_env_data.lib_deps32}	
 	${common_env_data.lib_ble32}	
 lib_ignore = 
-	Arduino-Temperature-Control-Library
 	mpu6050
 board = lolin_s3_mini 
 build_type = release
