+++ src/config_gravitymon.cpp
@@ -59,7 +59,6 @@ void GravitymonConfig::createJson(JsonObject& doc) {
   doc[CONFIG_FORMULA_DEVIATION] =
       serialized(String(this->getMaxFormulaCreationDeviation(), DECIMALS_TILT));
   doc[CONFIG_FORMULA_CALIBRATION_TEMP] = this->getDefaultCalibrationTemp();
-  doc[CONFIG_TEMPSENSOR_RESOLUTION] = this->getTempSensorResolution();
   doc[CONFIG_IGNORE_LOW_ANGLES] = this->isIgnoreLowAnges();
   doc[CONFIG_BLE_FORMAT] = getGravitymonBleFormat();
   doc[CONFIG_BATTERY_SAVING] = this->isBatterySaving();
@@ -126,8 +125,6 @@ void GravitymonConfig::parseJson(JsonObject& doc) {
   if (!doc[CONFIG_FORMULA_CALIBRATION_TEMP].isNull())
     this->SetDefaultCalibrationTemp(
         doc[CONFIG_FORMULA_CALIBRATION_TEMP].as<float>());
-  if (!doc[CONFIG_TEMPSENSOR_RESOLUTION].isNull())
-    this->setTempSensorResolution(doc[CONFIG_TEMPSENSOR_RESOLUTION].as<int>());
   if (!doc[CONFIG_IGNORE_LOW_ANGLES].isNull())
     setIgnoreLowAnges(doc[CONFIG_IGNORE_LOW_ANGLES].as<bool>());
   if (!doc[CONFIG_BLE_FORMAT].isNull())
