+++ src/tempsensor.cpp
@@ -21,15 +21,19 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
  */
-#if defined(GRAVITYMON)
+#if defined(GRAVITYMON) || defined(PRESSUREMON)
+
 #include <OneWire.h>
 #include <Wire.h>
 
 #include <config.hpp>
-#include <gyro.hpp>
 #include <log.hpp>
 #include <main.hpp>
 #include <tempsensor.hpp>
+
+#if defined(GRAVITYMON)
+#include <gyro.hpp>
+#endif // GRAVITYMON
 
 OneWire myOneWire(PIN_DS);
 DallasTemperature mySensors(&myOneWire);
@@ -59,6 +63,7 @@ void TempSensor::setup() {
 }
 
 void TempSensor::readSensor(bool useGyro) {
+#if defined(GRAVITYMON)
   if (useGyro) {
     // When using the gyro temperature only the first read value will be
     // accurate so we will use this for processing.
@@ -70,6 +75,7 @@ void TempSensor::readSensor(bool useGyro) {
 #endif
     return;
   }
+#endif // GRAVITYMON
 
   // If we dont have sensors just return 0
   if (!mySensors.getDS18Count()) {
@@ -95,6 +101,6 @@ void TempSensor::readSensor(bool useGyro) {
   }
 }
 
-#endif  // GRAVITYMON
+#endif  // GRAVITYMON || PRESSUREMON
 
 // EOF
